+++ src/lib.rs
@@ -9,6 +9,7 @@ pub mod cli;
 pub mod config;
 pub mod data;
 pub mod importer;
+pub mod keyboard;
 pub mod midi;
 pub mod model;
 pub mod music;